Add Remark Command
Add Remark Command is based on the tp tutorial from cs2103-2324s1's
github website in order to better understand how the existing code
base for the AB3 works. The summary of changes is as follows:

Firstly, we trace through the program and see that the entry point
for executing user commmands is via ui.MainWindow::executeCommand,
which gets the LogicManager object to obtain a CommandResult, which
is then shown the user as feedback.

For the remark command, LogicManager::execute parses the user input
using AddressBookParser::parseCommand, which makes use of
RemarkCommandParser::parse to return a valid RemarkCommand object,
before RemarkCommand::execute is called which updates the person in
the model with the new remark.

Next, LogicManager::execute saves the updated model into the hard
disk, before finally returning the CommandResult object, which is
passed back to the MainWindow object to show the feedback to the
user.